refactor(agent): split veryfront-cloud-agent-service into concern modules
Decompose the 1551-line monolith into five sibling modules under
src/agent/hosted/, each owning one concern:

- cloud-agent-paths.ts        — path/dir resolution, service naming, env
- cloud-agent-provider-bootstrap.ts — schema/auth/telemetry/bash defaults
- cloud-agent-config.ts       — context creation, agent-config resolution, steering
- cloud-agent-child-tools.ts  — MCP resolution, delegation, invoke-agent, skill/steering accessors
- cloud-agent-chat-execution.ts — chat execution preparation, runtime wiring

veryfront-cloud-agent-service.ts retains all 7 public function/const exports
and all public types as thin re-exports; no importer of the file changes.
Zero behavior change. All 26 tests pass.
